day 12

+++ src/helpers/line.rs
@@ -23,9 +23,9 @@ impl Iterator for LineIter {
                 return None;
             }
             let line_vec = (self.b.0 - self.a.0, self.b.1 - self.a.1);
-            let line_len = ((line_vec.0 * line_vec.0 + line_vec.1 * line_vec.1) as f32).sqrt();
-            let line_vec = (line_vec.0 as f32 / line_len, line_vec.1 as f32 / line_len);
-            let pos = ((current.0 - self.a.0) as f32, (current.1 - self.a.1) as f32);
+            let line_len = ((line_vec.0 * line_vec.0 + line_vec.1 * line_vec.1) as f64).sqrt();
+            let line_vec = (line_vec.0 as f64 / line_len, line_vec.1 as f64 / line_len);
+            let pos = ((current.0 - self.a.0) as f64, (current.1 - self.a.1) as f64);
             let delta = *[
                 (line_vec.0.signum() as i32, 0),
                 (0, line_vec.1.signum() as i32),
@@ -33,7 +33,7 @@ impl Iterator for LineIter {
             ]
             .iter()
             .min_by_key(|delta| {
-                let new_pos = (pos.0 + delta.0 as f32, pos.1 + delta.1 as f32);
+                let new_pos = (pos.0 + delta.0 as f64, pos.1 + delta.1 as f64);
                 let dot = new_pos.0 * line_vec.0 + new_pos.1 * line_vec.1;
                 let v = (new_pos.0 - line_vec.0 * dot, new_pos.1 - line_vec.1 * dot);
                 OrderedFloat(v.0 * v.0 + v.1 * v.1)
